Comments added

+++ CampaignCodeGenerator/Models/CampaignCodeModel.cs
@@ -1,8 +1,9 @@
 ﻿namespace CampaignCodeGenerator.Models
 {
+    //If the user wants to get the codes with their Ids, this model will be used.
     public class CampaignCodeModel
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Code { get; set; }
     }
 }
